Move onboarding config into separate package
